Move BuiltinFunc to _internal_types
